requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pytz==2024.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
